safari-600.7-branch/LayoutTests/fast/selectors/querySelector-long-child-backtracking-expected.txt

Test backtracking with multiple child selector relations.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.querySelectorAll("ul>li>div>pre span").length is 3
PASS document.querySelectorAll("ul>li>div>pre span")[0].id is "target1"
PASS document.querySelectorAll("ul>li>div>pre span")[1].id is "target2"
PASS document.querySelectorAll("ul>li>div>pre span")[2].id is "target3"
PASS successfullyParsed is true

TEST COMPLETE
